move reaction component prototype modules out of node_modules so that they will be included in npm releases. fix #93
